# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ComponentHealth": {
        "properties": {
          "contentStore": {
            "const": "ready",
            "title": "Contentstore",
            "type": "string"
          },
          "database": {
            "const": "ready",
            "title": "Database",
            "type": "string"
          },
          "worker": {
            "const": "ready",
            "title": "Worker",
            "type": "string"
          }
        },
        "required": [
          "database",
          "contentStore",
          "worker"
        ],
        "title": "ComponentHealth",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "components": {
            "$ref": "#/components/schemas/ComponentHealth"
          },
          "schemaVersion": {
            "const": 1,
            "default": 1,
            "title": "Schemaversion",
            "type": "integer"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "components"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "SessionBootstrapRequest": {
        "properties": {
          "officeHost": {
            "const": "Word",
            "title": "Officehost",
            "type": "string"
          },
          "officePlatform": {
            "enum": [
              "PC",
              "Mac"
            ],
            "title": "Officeplatform",
            "type": "string"
          },
          "wordApi13Supported": {
            "const": true,
            "title": "Wordapi13Supported",
            "type": "boolean"
          }
        },
        "required": [
          "officeHost",
          "officePlatform",
          "wordApi13Supported"
        ],
        "title": "SessionBootstrapRequest",
        "type": "object"
      },
      "SessionBootstrapResponse": {
        "properties": {
          "csrfToken": {
            "title": "Csrftoken",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          }
        },
        "required": [
          "csrfToken",
          "expiresAt"
        ],
        "title": "SessionBootstrapResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Word Researcher Local Companion",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Health"
      }
    },
    "/api/v1/session/bootstrap": {
      "post": {
        "operationId": "bootstrapLocalSession",
        "parameters": [
          {
            "in": "header",
            "name": "x-bootstrap-csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Bootstrap-Csrf"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionBootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBootstrapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap Session"
      }
    }
  }
}